Make application cache more resilient to startup situations
When starting gate and clouddriver at the same time, gate would sometimes fail and lose the "ticking" that refreshes its cache. This patch moves the ticking subscription out of the constructor call and into a @PostConstruct method. It also wraps the tick() method with a try-catch clause so that it will keep on ticking despite exceptions.