feat(authN): Restrict github auth by organization membership
This adds the ability to restrict github authentication to only users
belonging to a certain organization. If there's a
userInfoRequirements.organization defined in the gate config and an
organizations_url is returned in the user info for an oauth2 request we
make a second request to the organizations_url and merge that data into
the map we return as the initial user object. The
hasAllUserInfoRequirements() function can then filter based on
organization membership. That works a lot like the existing filters, but
because the organizations come back as an array a simple static match or
regex wasn't enough. The user does have to have their membership in the
organization set to "public" in order for this check to be able to see
it. This is a proposed fix for
https://github.com/spinnaker/fiat/issues/188