feat(context): replace RequestContext with AuthenticatedRequest
This makes gate's RequestContext a readonly proxy to kork's AuthenticatedRequest. This is to avoid the situation where
RequestContext and AuthenticatedRequest provide an inconsistent view of the context or lose it across thread boundaries.

This is a relatively big change, values in AuthenticatedRequest get propagated across threads and to downstream services
in the form of X-SPINNAKER headers. So with this change, gate will now start emitting X-SPINNAKER-APPLICATION headers for
instance (until now, only orca did that).