feat(serviceAccounts): filter service accounts by application
enables a more meaningful list of service accounts when selecting a run as user

This feature can be disabled by setting `fiat.service-accounts.filter.enabled=false`.

The required authorizations for the service account on the application can be customized,
by default they are either of WRITE or EXECUTE, but can be set as
`fiat.service-accounts.filter.match-authorizations=READ,WRITE,EXECUTE`

In the event that no service accounts match the specific application, the user's full
list of service accounts is returned as a fallback.