fix(etag): produce weak ETag
follow up from: https://github.com/spinnaker/gate/pull/1123

We noticed that `ETag` filter somehow makes it so that requests from `gate` are no longer gzipped.
Removing the filter caused the issue that responses no longer contained `Content-Length` header which means that the responses were chunked.
This is a pretty big difference and a breaking change for clients that might not be able to handle chunked responses.

This change sets the `ETag` to be weak which doesn't circumvent gzip but also sets the `content-length` and is thus not chunked.