feat(api): An `ExtensionPoint` that a plugin to expose endpoints under a common path
This implementation is similar to that of the `ProxyController` and allows for
new endpoints to be exposed under the `/extensions/{id}/**` path.

I've added `ApiExtensionController` into the `gate-proxy` artifact as there is
some overlap in behavior with the existing `ProxyController`. I expect the
`ProxyController` to receive a similar extension point, at which time it  might
be appropriate to rename the artifact.

This PR also bumps `kotlin` from 1.3.21 to 1.3.71.