fix(build): support nested composite builds
Following the composte build flags convention, this change
prevents including builds in this project if a flag exists
that would indicate this project is being included in a
composite build already.

If `<projectName>Composite` exists and is set to `true`,
then this project won't include any builds. The project
that has imported this project is responsible for build
inclusion.