fix(web): Default `Content-Type` within `DataController`
Prior to this PR, the `Content-Type` was defaulting to `text/html` which
rendered correctly in the browser but appeared to cause problems with `deck`.

Given a narrow set of use-cases, these endpoints now default to returning
`application/json` and use the `OutputStream` from `HttpServletResponse`
rather than re-creating a new one and copying bytes.