fix(deps): adds an interface for request identity extractors
This allows for checks based on some identity to occur
before triggering the auth flow for cases such as rate
limiting where we want to reject the request doing the
least amount of work possible.

Previously this was done by directly depending on x509.
This change introduces an interface in gate-core  and an
implementation of that interface in the gate-x509 module.

This was mostly just shuffling existing code around more
than adding anything specifically new but it gets back to
the auth implementations being optional dependencies.